refactor: add dnd hooks and shared constants
