+++ src/features/kanban/board/components/modals/NewBoardModal.tsx
@@ -1,8 +1,9 @@
+import { BoardSchema } from 'features/kanban/board/board.schema'
+import { useAddBoard } from 'features/kanban/board/hooks'
+
 import { Button, Field, Modal } from 'components/ui'
 import { useAppForm } from 'hooks'
 
-import { BoardSchema } from '../../board.schema'
-import { useAddBoard } from '../../hooks'
 import { RadioInputBgImages } from './RadioInputBgImages'
 import { RadioInputIcons } from './RadioInputIcons'
