+++ src/features/kanban/card/card.types.ts
@@ -1,4 +1,4 @@
-import type { Priority } from './card.constants'
+import type { Priority } from '../shared/constants'
 
 export type Card = {
   id: string
