+++ src/features/kanban/card/components/modals/NewCardModal.tsx
@@ -1,9 +1,9 @@
+import { CardSchema } from 'features/kanban/card/card.schema'
 import { useAddCard } from 'features/kanban/card/hooks'
 
 import { Button, Field, Modal } from 'components/ui'
 import { useAppForm } from 'hooks'
 
-import { CardSchema } from '../../card.schema'
 import { DatePicker } from '../ui'
 import { ModalDescription } from './ModalDescription'
 import { ModalPriorities } from './ModalPriorities'
