+++ src/features/kanban/dnd/hooks/useDragAndDrop.ts
@@ -1,0 +1,13 @@
+import { useContext } from 'react'
+
+import { DragAndDropContext } from '../dnd.context'
+
+export const useDragAndDrop = () => {
+  const context = useContext(DragAndDropContext)
+
+  if (!context) {
+    throw new Error('useDragAndDrop must be used within a DragAndDropProvider')
+  }
+
+  return context
+}
