+++ src/features/kanban/filters/components/DeadlineFilter.tsx
@@ -1,9 +1,8 @@
 import { Root } from '@radix-ui/react-radio-group'
 
 import { useCardFilters } from 'features/kanban/card/hooks'
-import { RadioInput } from 'features/kanban/shared/components/ui'
-
-import { DEADLINES } from '../filters.constants'
+import { RadioInput } from 'features/kanban/shared/components'
+import { DEADLINES } from 'features/kanban/shared/constants'
 
 export const DeadlineFilter = () => {
   const { setSearchParams, cardDeadline } = useCardFilters()
