--- src/features/kanban/filters/filters.constants.ts
@@ -1,3 +1,0 @@
-export const DEADLINES = ['Upcoming', 'Overdue', 'Far Future'] as const
-
-export type Deadline = (typeof DEADLINES)[number]
